+++ R/fn.R
@@ -422,7 +422,7 @@ as_function <- function(x, env = caller_env()) {
       abort("Can't convert a two-sided formula to a function")
     }
 
-    args <- exprs(... = , .x = ..1, .y = ..2, . = ..1)
+    args <- list(... = missing_arg(), .x = quote(..1), .y = quote(..2), . = quote(..1))
     fn <- new_function(args, f_rhs(x), f_env(x))
     fn <- structure(fn, class = c("rlang_lambda_function", "function"))
     return(fn)