+++ tests/testthat/test-types.R
@@ -18,19 +18,6 @@ test_that("can bypass string serialisation", {
   expect_identical(Encoding(bar[[2]]), "latin1")
 })
 
-test_that("lang_type_of() returns correct lang subtype", {
-  expect_identical(lang_type_of(quote(foo())), "named")
-  expect_identical(lang_type_of(quote(foo::bar())), "namespaced")
-  expect_identical(lang_type_of(quote(foo@bar())), "recursive")
-
-  lang <- quote(foo())
-  node_poke_car(lang, 10)
-  expect_error(lang_type_of(lang), "corrupt")
-
-  node_poke_car(lang, base::list)
-  expect_identical(lang_type_of(lang), "inlined")
-})
-
 test_that("types are friendly", {
   expect_identical(friendly_type("character"), "a character vector")
   expect_identical(friendly_type("integer"), "an integer vector")